Update SASSCAL_WebSAPI.Rmd
This RMarkdown is meant to demonstrate the proof of concept that R facilitates end users to take control of the online SASSCAL weather data.  It also allows visualization of the data using tables and maps. More functionalities will be added to enhance usability in the form of a complete dashboard and analytics system.